Single png render in export rather than also in gif.
